+++ empire/client/src/utils/date_util.py
@@ -17,7 +17,3 @@ def humanize_datetime(iso_string: str | None = None):
     local_str = parsed.astimezone().strftime("%Y-%m-%d %H:%M:%S %Z")
 
     return f"{local_str} ({humanize.naturaltime(datetime.now(UTC) - parsed)})"
-
-
-def get_utc_now():
-    return datetime.now(UTC).strftime("%Y-%m-%d %H:%M:%S")